got delete working

+++ app/api/table_routes.py
@@ -76,7 +76,7 @@ def edit(id):
 
             return edited_gametable.to_dict()
         else:
-            return 'You do not have access to edit this comment!'
+            return 'You do not have access to edit this Game Table!'
     else:
 
         return form.errors
@@ -89,10 +89,14 @@ def delete(id):
     GameTable DELETE route.
     '''
     gametable_to_delete = GameTable.query.filter(GameTable.id == id).first()
+    userId = current_user.get_id()
 
     if not gametable_to_delete:
         return 'Nothing to delete'
     else:
-        db.session.delete(gametable_to_delete)
-        db.session.commit()
-        return gametable_to_delete.to_dict()
+        if int(userId) == int(gametable_to_delete.tableCreator):
+            db.session.delete(gametable_to_delete)
+            db.session.commit()
+            return gametable_to_delete.to_dict()
+        else:
+            return 'You do not have access to delete this Game Table!'
